fix run.add_comm
previously, CRS and CRE as well as ref were not really implemented right, so the highlights were all messed up. This commit fixes that, so now the add_comm method works as expected. It also adds some helpful functions that will come in handy for adding comments over multiple runs