DEV-2853: fix all_runs methods for Ins and Del
Issue here we are finding is that these methods are excluding runs that have a run as an ancestor, but that does not really matter for what these methods are for. We want to know: inside THIS Ins or Del, give me all of the runs. It looks like I copied this method from one in Paragraph, which was written by somebody else.